Move pytest fixtures out to conftest for general availability
